Changed Celery to APSscheduler
