+++ backend/.env
@@ -3,20 +3,22 @@ SECRET_KEY = "SUPER_SECRET_KEY" #change in production
 JWT_ALGORITHM = "HS256"
 ACCES_JWT_EXPIRY_SECONDS = "3600"
 REFRESH_JWT_EXPIRY_SECONDS = "36000"
-
-POST_RATING_EXPIRATION = "5"
-
 QUERY_PARAM_MAX_L = "500"
 
-FEED_MAX_POSTS_LOAD = "50"
+POST_RATING_EXPIRATION = "60"
 
+# Post actions cost
+VIEW = "1"
+LIKE = "5"
+REPLY = "5"
+REPOST = "8"
+
+FEED_MAX_POSTS_LOAD = "50"
 # Variables down here need to sum up to FEED_MAX_POSTS_LOADED
 N_MAX_HISTORY_POSTS_SHOW = "25"
 N_MAX_FRESH_POSTS_TO_MIX = "25"
-
 # Preferably be same as FEED_MAX_POSTS_LOADED
 MAX_FOLLOWED_POSTS_TO_SHOW = "50"
-
 
 DATETIME_BASE_FORMAT = "%d/%m/%Y, %H:%M:%S"
 
@@ -25,7 +27,7 @@ HISTORY_POSTS_TO_TAKE_INTO_RELATED = "30"
 PASSWORD_MIN_L = "8"
 PASSWORD_MAX_L = "32"
 
-# CELERY | Call in backend directory
+# CELERY | Call in /backend directory
 # celery -A post_popularity_rate_celery.popularity_rate  worker --loglevel=info -P gevent
 # celery -A post_popularity_rate_celery.celery_main beat -l info
