+++ backend/.gitignore
@@ -1,5 +1,2 @@
 /.pytest_cache
 __pycache__/
-*.bak
-*.dat
-*.dir
