--- backend/celerybeat-schedule.bak
@@ -1,4 +1,0 @@
-'entries', (2048, 521)
-'__version__', (512, 20)
-'tz', (1024, 4)
-'utc_enabled', (1536, 4)
