+++ backend/main.py
@@ -13,8 +13,12 @@ import redis.asyncio as async_redis
 import uvicorn
 from os import getenv
 from dotenv import load_dotenv
+from post_popularity_rate_task.popularity_rate import update_post_rates
+
+from post_popularity_rate_task.popularity_rate import scheduler
 
 load_dotenv()
+POST_RATING_EXPIRATION = int(getenv("POST_RATING_EXPIRATION"))
 
 async def drop_redis() -> None:
     client = async_redis.Redis(
@@ -38,10 +42,22 @@ async def sync_chroma_postgres_data() -> None:
 # On app startup. https://fastapi.tiangolo.com/advanced/events/#lifespan
 @asynccontextmanager
 async def lifespan(app: FastAPI):
+
     # await drop_all(engine=engine, Base=Base)    
     await initialize_models(engine=engine, Base=Base)
     await sync_chroma_postgres_data()
     # await drop_redis()
+
+    try:
+        scheduler.add_job(
+            update_post_rates,
+            "interval",
+            seconds=POST_RATING_EXPIRATION
+        )
+        scheduler.start()
+    except Exception as e:
+        scheduler.shutdown()
+        raise e("Scheduler initializtion failed")
     yield
 
 
@@ -59,7 +75,6 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
 
 app.include_router(auth_router.auth)
 app.include_router(social_router.social)
