--- backend/post_popularity_rate_celery/.gitignore
@@ -1,1 +1,0 @@
-__pycaache__/
