--- backend/post_popularity_rate_celery/celery_main.py
@@ -1,5 +1,0 @@
-from celery import Celery
-
-celery_app = Celery("tasks", broker="redis://localhost")
-
-from post_popularity_rate_celery import popularity_rate
