+++ backend/post_popularity_rate_task/.gitignore
@@ -1,0 +1,1 @@
+__pycaache__/
